+++ src/test/java/nextstep/subway/path/domain/PathTest.java
@@ -3,8 +3,6 @@ package nextstep.subway.path.domain;
 import java.util.Arrays;
 import java.util.List;
 import nextstep.subway.line.domain.Line;
-import nextstep.subway.line.domain.LineTest;
-import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.StationTest;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
@@ -19,7 +17,7 @@ public class PathTest {
     private Line 이호선;
     private Line 일호선;
 
-    void setUp(){
+    void setUp() {
         이호선 = new Line("이호선", "green", StationTest.STATION_4, StationTest.STATION_5, 10);
         이호선.addStation(StationTest.STATION_3, StationTest.STATION_5, 4);
         일호선 = new Line("일호선", "navy", StationTest.STATION_4, StationTest.STATION_5, 2);
@@ -41,7 +39,7 @@ public class PathTest {
     }
 
     @Test
-    void create_path(){
+    void create_path() {
         //given
         setUp();
 
@@ -55,10 +53,10 @@ public class PathTest {
     }
 
     @Test
-    void find_shortest_path(){
+    void find_shortest_path() {
         //given
         setUp();
-        double expected = 2.00;
+        int expected = 2;
         Path path = Path.of(Arrays.asList(이호선, 일호선));
 
         //when
